feat: setup router and implement homeView route
