+++ src/main.js
@@ -1,6 +1,10 @@
 import './assets/main.css'
+import router from './router';
 
 import { createApp } from 'vue'
 import App from './App.vue'
 
-createApp(App).mount('#app')
+const app = createApp(App);
+
+app.use(router);
+app.mount('#app');
