+++ src/router/index.js
@@ -1,0 +1,18 @@
+import { createRouter, createWebHistory } from "vue-router";
+
+import HomeView from "@/views/HomeView.vue";
+
+const routes = [
+    {
+        path: '/',
+        name: 'home',
+        component: HomeView
+    }
+]
+
+const router = createRouter({
+    history: createWebHistory(import.meta.env.BASE_URL),
+    routes
+});
+
+export default router;
